fix: <select> always remembers selected state (#32)
Simplified the implementation quite a bit. Now upon hitting <Enter> on a
select, we show a custom select on top of it, which options just set up
the select.value on the REAL select.

This means the "customSelect" only exists when the "real select" is
programatically expanded. IE the real select is the source of truth,
which makes everything so much more simpler.